oop and class design chapter

+++ oop/src/main/java/ocp/oop/example4/Example1.java
@@ -6,15 +6,16 @@ public class Example1 {
 
     /**
      * Overloading
-     * should know, which method called
+     * should know, which method called ?
      */
 
     A a = new A();
+
     a.a(10); // call a(int x)
     a.a(12.3); // call a(double x)
-    a.a(13L); // call a(double x)
-
+    a.a((Long)13L); // call a(long x)
     a.a((int)18.0); // call a(int x)
+
 
     // a.a(null); // CF: Ambiguous method call. Both  a(String s) and a(Double d) match
     a.a((Double) null); // to avoid this problem explicit cast is needed
